Add integration tests for GCP, fix linter errors
